requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
Authlib==1.3.0
backoff==2.2.1
bcrypt==5.0.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorbrewer==0.3
colour==0.1.5
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
ecdsa==0.19.1
fastapi==0.104.1
fonttools==4.60.1
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
google-cloud-secret-manager==2.16.0
google-cloud-storage==2.10.0
google-cloud-texttospeech==2.14.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
igraph==1.0.0
importlib_resources==6.5.2
itsdangerous==2.1.2
jiter==0.12.0
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mido==1.3.3
networkx==3.6.1
numpy==2.3.4
openai==1.54.0
packaging==24.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
pretty_midi==0.2.11
proto-plus==1.27.0
protobuf==4.25.8
py4cytoscape==1.12.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyodbc==5.0.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tabulate==0.9.0
texttable==1.7.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
